snapcast: show snapweb placeholder
snapserver has a built-in webserver which is able to host a web player:
snapweb. [1]
It is accessible at http://snapserver:1780/ but with enabled "doc_root"
setting and with missing directory the browser shows an empty page
with a message: "The resource '/' was not found."

Fix this with disabling the "doc_root" setting in "/etc/snapserver.conf"
This "doc_root" was enabled and set to "/usr/share/snapserver/snapweb"
in upstream commit fcbe40585875 ("Install snapweb") [2], back in 2020
and released in v0.21.0 of Snapcast.

This was reported in an upstream discussion by Robert Muth. [3]

[1]: https://github.com/badaix/snapweb
[2]: https://github.com/badaix/snapcast/commit/fcbe405858752457a67006eeff4f0da60e8fbc89
[3]: https://github.com/badaix/snapcast/discussions/1417#discussioncomment-14480178

Fixes: eeb8d131fc9 ("snapcast: add package snapserver and snapclient")
Signed-off-by: Szabolcs Hubai <szab.hu@gmail.com>